Adds supacec experimental package
This package will listen for new input devices and will send a cec signal that will take over the display. So, when you turn on your controller, the display switches over. A bit rudimentary but it works, hoping to find it a proper home where some people might use it.